+++ pywbemtools/pywbemcli/__init__.py
@@ -32,6 +32,7 @@ from ._cmd_server import *          # noqa: F403,F401
 from ._cmd_connection import *      # noqa: F403,F401
 from ._cmd_profile import *         # noqa: F403,F401
 from ._cmd_statistics import *      # noqa: F403,F401
+from ._cmd_listener import *        # noqa: F403,F401
 from ._context_obj import *         # noqa: F403,F401
 from ._connection_repository import *   # noqa: F403,F401
 from .pywbemcli import *            # noqa: F403,F401